+++ packages/showcases/svelte/stories/components/actions/VtmnLink/VtmnLink.stories.svelte
@@ -1,7 +1,10 @@
 <script>
   import { Meta, Template, Story } from '@storybook/addon-svelte-csf';
   import { VtmnLink } from '@vtmn/svelte';
-  import { parameters } from '@vtmn/showcase-core/csf/components/actions/link.csf';
+  import {
+    parameters,
+    argTypes,
+  } from '@vtmn/showcase-core/csf/components/actions/link.csf';
 
   const linkArgs = {
     href: 'https://decathlon.design',
@@ -13,6 +16,7 @@
   title="Components / Actions / VtmnLink"
   component={VtmnLink}
   {parameters}
+  {argTypes}
 />
 
 <Template let:args>