+++ packages/sources/svelte/src/components/index.js
@@ -8,17 +8,31 @@ export { default as VtmnTextInput } from './forms/VtmnTextInput/VtmnTextInput.sv
 // Indicators
 export { default as VtmnBadge } from './indicators/VtmnBadge/VtmnBadge.svelte';
 export { default as VtmnPrice } from './indicators/VtmnPrice/VtmnPrice.svelte';
+export { default as VtmnRating } from './indicators/VtmnRating/VtmnRating.svelte';
 export { default as VtmnTag } from './indicators/VtmnTag/VtmnTag.svelte';
 
 // Navigation
+export { default as VtmnSearch } from './navigation/VtmnSearch/VtmnSearch.svelte';
 
 // Overlays
+export { default as VtmnAlert } from './overlays/VtmnAlert/VtmnAlert.svelte';
+export { vtmnAlertStore } from './overlays/VtmnAlert/vtmnAlertStore';
+export { default as VtmnModal } from './overlays/VtmnModal/VtmnModal.svelte';
 export { default as VtmnPopover } from './overlays/VtmnPopover/VtmnPopover.svelte';
+export { default as VtmnSnackbar } from './overlays/VtmnSnackbar/VtmnSnackbar.svelte';
+export { vtmnSnackbarStore } from './overlays/VtmnSnackbar/vtmnSnackbarStore';
+export { default as VtmnToast } from './overlays/VtmnToast/VtmnToast.svelte';
+export { vtmnToastStore } from './overlays/VtmnToast/vtmnToastStore';
 
 // Selection controls
 export { default as VtmnCheckbox } from './selection-controls/VtmnCheckbox/VtmnCheckbox.svelte';
+export { default as VtmnChip } from './selection-controls/VtmnChip/VtmnChip.svelte';
 export { default as VtmnToggle } from './selection-controls/VtmnToggle/VtmnToggle.svelte';
 
 // Structure
+export { default as VtmnAccordion } from './structure/VtmnAccordion/VtmnAccordion.svelte';
+export { default as VtmnCard } from './structure/VtmnCard/VtmnCard.svelte';
+export { default as VtmnDivider } from './structure/VtmnDivider/VtmnDivider.svelte';
 export { default as VtmnList } from './structure/VtmnList/VtmnList.svelte';
 export { default as VtmnListItem } from './structure/VtmnListItem/VtmnListItem.svelte';
+export { default as VtmnSkeleton } from './structure/VtmnSkeleton/VtmnSkeleton.svelte';