+++ packages/sources/svelte/src/components/structure/VtmnSkeleton/enums.js
@@ -1,0 +1,7 @@
+/**
+ * @type {string} Custom classes to apply to the component for the nature of the shape.
+ */
+export const VTMN_SKELETON_SHAPE = {
+  LINE: 'line',
+  AVATAR: 'avatar',
+};