+++ packages/sources/vue/src/components/forms/VtmnSelect/VtmnSelect.vue
@@ -5,6 +5,7 @@ import { VtmnSelectOption } from './types';
 
 export default /*#__PURE__*/ defineComponent({
   name: 'VtmnSelect',
+  inheritAttrs: false,
   props: {
     modelValue: {
       type: [String, Number, Boolean] as PropType<string | number | boolean>,
@@ -67,7 +68,7 @@ export default /*#__PURE__*/ defineComponent({
 </script>
 
 <template>
-  <div className="vtmn-select_container">
+  <div class="vtmn-select_container">
     <label :v-if="labelText" class="vtmn-text-input_label" :for="identifier">
       {{ labelText }}
     </label>