+++ packages/sources/css/presets/tailwind/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.33.80](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.79...@vtmn/css-tailwind-preset@0.33.80) (2022-03-18)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.79](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.78...@vtmn/css-tailwind-preset@0.33.79) (2022-03-11)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.78](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.77...@vtmn/css-tailwind-preset@0.33.78) (2022-03-11)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.77](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.76...@vtmn/css-tailwind-preset@0.33.77) (2022-03-09)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.76](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.75...@vtmn/css-tailwind-preset@0.33.76) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
 ## [0.33.75](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.74...@vtmn/css-tailwind-preset@0.33.75) (2022-03-04)
 
 **Note:** Version bump only for package @vtmn/css-tailwind-preset