+++ packages/sources/css/src/components/actions/link/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.15](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-link@0.6.14...@vtmn/css-link@0.6.15) (2022-03-11)
+
+**Note:** Version bump only for package @vtmn/css-link
+
+
+
+
+
+## [0.6.14](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-link@0.6.13...@vtmn/css-link@0.6.14) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css-link
+
+
+
+
+
 ## [0.6.13](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-link@0.6.12...@vtmn/css-link@0.6.13) (2022-03-03)
 
 **Note:** Version bump only for package @vtmn/css-link