+++ packages/sources/css/src/components/indicators/rating/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.13](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-rating@0.3.12...@vtmn/css-rating@0.3.13) (2022-03-11)
+
+**Note:** Version bump only for package @vtmn/css-rating
+
+
+
+
+
+## [0.3.12](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-rating@0.3.11...@vtmn/css-rating@0.3.12) (2022-03-09)
+
+**Note:** Version bump only for package @vtmn/css-rating
+
+
+
+
+
+## [0.3.11](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-rating@0.3.10...@vtmn/css-rating@0.3.11) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css-rating
+
+
+
+
+
 ## [0.3.10](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-rating@0.3.9...@vtmn/css-rating@0.3.10) (2022-03-03)
 
 **Note:** Version bump only for package @vtmn/css-rating