+++ packages/sources/css/src/components/overlays/snackbar/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.13](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-snackbar@0.5.12...@vtmn/css-snackbar@0.5.13) (2022-03-11)
+
+**Note:** Version bump only for package @vtmn/css-snackbar
+
+
+
+
+
+## [0.5.12](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-snackbar@0.5.11...@vtmn/css-snackbar@0.5.12) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css-snackbar
+
+
+
+
+
 ## [0.5.11](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-snackbar@0.5.10...@vtmn/css-snackbar@0.5.11) (2022-03-03)
 
 **Note:** Version bump only for package @vtmn/css-snackbar