+++ packages/sources/css/src/components/selection-controls/toggle/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.12](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-toggle@0.6.11...@vtmn/css-toggle@0.6.12) (2022-03-11)
+
+**Note:** Version bump only for package @vtmn/css-toggle
+
+
+
+
+
+## [0.6.11](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-toggle@0.6.10...@vtmn/css-toggle@0.6.11) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css-toggle
+
+
+
+
+
 ## [0.6.10](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-toggle@0.6.9...@vtmn/css-toggle@0.6.10) (2022-03-03)
 
 **Note:** Version bump only for package @vtmn/css-toggle