+++ packages/sources/css/src/components/structure/accordion/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.13](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-accordion@0.2.12...@vtmn/css-accordion@0.2.13) (2022-03-11)
+
+**Note:** Version bump only for package @vtmn/css-accordion
+
+
+
+
+
+## [0.2.12](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-accordion@0.2.11...@vtmn/css-accordion@0.2.12) (2022-03-09)
+
+**Note:** Version bump only for package @vtmn/css-accordion
+
+
+
+
+
+## [0.2.11](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-accordion@0.2.10...@vtmn/css-accordion@0.2.11) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css-accordion
+
+
+
+
+
 ## [0.2.10](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-accordion@0.2.9...@vtmn/css-accordion@0.2.10) (2022-03-03)
 
 **Note:** Version bump only for package @vtmn/css-accordion